Replace height: 100% with min-height: 100vh
A passing convo revealed to me the existence of "min-height: 100vh" on
the <body> as an alternative to the usual "height: 100%" on the <html>
and <body>.

Searching online led me to [1], which shows that 100vh is better because it
doesn't force the page height to be the window height, but instead
allows the page height to conform to the page content's actual height.

Thanks to @adri326 for mentioning it.

[1]: https://greggod.medium.com/css-do-not-put-height-100-on-the-body-html-e36bda3551b3